third_party/blink/web_tests/platform/generic/fast/frames/sandboxed-iframe-storage-expected.txt

This test verifies that a sandboxed iframe does not have access to any session storage, local storage or database.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Most of the PASS messages appear in frames, presented by DumpRenderTree after the main window.

 
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS window.openDatabase('SandboxedIframeStorageDisallowed', '1.0', '', 1) threw exception "SecurityError: Failed to execute 'openDatabase' on 'Window': Access to the WebDatabase API is denied in this context.".
PASS window.indexedDB.open('foo') threw exception "SecurityError: Failed to execute 'open' on 'IDBFactory': access to the Indexed Database API is denied in this context.".
PASS window.indexedDB.deleteDatabase('foo') threw exception "SecurityError: Failed to execute 'deleteDatabase' on 'IDBFactory': access to the Indexed Database API is denied in this context.".
PASS window.localStorage threw exception "SecurityError: Failed to read the 'localStorage' property from 'Window': The document is sandboxed and lacks the 'allow-same-origin' flag.".
PASS window.sessionStorage threw exception "SecurityError: Failed to read the 'sessionStorage' property from 'Window': The document is sandboxed and lacks the 'allow-same-origin' flag.".
PASS document.cookie threw exception "SecurityError: Failed to read the 'cookie' property from 'Document': The document is sandboxed and lacks the 'allow-same-origin' flag.".


--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
PASS successfullyParsed is true

TEST COMPLETE

PASS window.openDatabase('SandboxedIframeStorageAllowed', '1.0', '', 1) != null is true
PASS window.localStorage != null is true
PASS window.sessionStorage != null is true
PASS document.cookie != null is true
